+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Changelog
 a) [TASK][!!!] Instead whole wp-cli/wp-cli load only wp-cli/search-replace-command.
 b) [TASK] Update deployer-extended-* packages.
 c) [FEATURE] Add sourcebroker/deployer-loader
+d) [TASK] Add .gitignore with /vendors so vendors can be installed while developing ext.
+
 
 0.3.1
 ~~~~~