+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ a) [TASK][!!!] Instead whole wp-cli/wp-cli load only wp-cli/search-replace-comma
 b) [TASK] Update deployer-extended-* packages.
 c) [FEATURE] Add sourcebroker/deployer-loader
 d) [TASK] Add .gitignore with /vendors so vendors can be installed while developing ext.
-
+e) [TASK] Move "db:import:post_command:wp_domains.php" to proper folder.
 
 0.3.1
 ~~~~~