+++ CHANGELOG.rst
@@ -11,7 +11,7 @@ c) [FEATURE] Add sourcebroker/deployer-loader
 d) [TASK] Add .gitignore with /vendors so vendors can be installed while developing ext.
 e) [TASK] Move "db:import:post_command:wp_domains.php" to proper folder.
 f) [TASK] Use Loader to load recipes from deployer-extended-*
-
+g) [TASK] Update docs.
 
 0.3.1
 ~~~~~