+++ CHANGELOG.rst
@@ -12,6 +12,8 @@ d) [TASK] Add .gitignore with /vendors so vendors can be installed while develop
 e) [TASK] Move "db:import:post_command:wp_domains.php" to proper folder.
 f) [TASK] Use Loader to load recipes from deployer-extended-*
 g) [TASK] Update docs.
+h) [TASK] In deploy add task do backup of database with rotation "last 10".
+
 
 0.3.1
 ~~~~~